+++ src/pages/HomePage.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import DownloadView from 'src/components/Pages/HOME/DownloadView.vue';
-import ProductIntro from 'src/components/Pages/HOME/ProductIntro.vue';
+import ProductIntroView from 'src/components/Pages/HOME/ProductIntroView.vue';
 import TitleView from 'src/components/Pages/HOME/TitleView.vue';
 import MultiPlayView from 'src/components/Pages/HOME/MultiPlayView.vue';
 import ImportWorldView from 'src/components/Pages/HOME/ImportWorldView.vue';
-import ServerCard from 'src/components/Pages/HOME/ServerCard.vue';
+import ServersView from 'src/components/Pages/HOME/ServersView.vue';
 import PortMappingView from 'src/components/Pages/HOME/PortMappingView.vue';
 import SummaryView from 'src/components/Pages/HOME/SummaryView.vue';
 </script>
@@ -28,9 +28,9 @@ import SummaryView from 'src/components/Pages/HOME/SummaryView.vue';
     <q-separator color="primary" size="3px" class="line" />
 
     <div class="mainField">
-      <ProductIntro />
+      <ProductIntroView />
       <MultiPlayView />
-      <ServerCard />
+      <ServersView />
       <PortMappingView />
       <ImportWorldView />
       <SummaryView />
